move test file to apt. location
